+++ database/migrations/2017_08_19_023022_create_developments_table.php
@@ -15,8 +15,8 @@ class CreateDevelopmentsTable extends Migration
     {
         Schema::create('developments', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('project_id');
-            $table->integer('status');
+            $table->unsignedInteger('project_id');
+            $table->unsignedTinyInteger('status');
             $table->timestamps();
             $table->softDeletes();
         });
